requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Faker==3.0.1
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
text-unidecode==1.3
twilio==6.48.0
urllib3==1.26.2
Werkzeug==1.0.1
